Breakout the drone dynamics from the 6dof (interfaces).
