+++ tools/tests/non_linear_sim/test_six_dof_model.py
@@ -1,6 +1,6 @@
 import pytest
 
-from non_linear_sim.non_linear_model import *
+from non_linear_sim.six_dof_model import *
 
 STATE_NON_ZERO = State(pos=1, vel=0.1, acc=0.01)
 
@@ -15,30 +15,30 @@ STATES_NON_ZERO = States(
 
 
 @pytest.fixture
-def nlm_zero_init():
-    yield NonLinearModel()
+def model_zero_init():
+    yield SixDofModel()
 
 
-class TestNonLinearModel:
+class TestSixDofModel:
 
     def test_init_without_arg(self):
-        nlm = NonLinearModel()
-        assert nlm.get_states() == STATES_ZERO
+        model = SixDofModel()
+        assert model.get_states() == STATES_ZERO
 
     @pytest.mark.parametrize("states", [STATES_ZERO, STATES_NON_ZERO])
     def test_init_with_arg(self, states):
-        nlm = NonLinearModel(states)
-        assert nlm.get_states() == states
+        model = SixDofModel(states)
+        assert model.get_states() == states
 
     def test_step(self):
         # TODO: Test step with various scenarios (zero input, const acc, step acc, back and fourth, ...)
         pass
 
-    def test_reset_without_arg(self, nlm_zero_init):
-        nlm_zero_init.reset()
-        assert nlm_zero_init.get_states() == STATES_ZERO
+    def test_reset_without_arg(self, model_zero_init):
+        model_zero_init.reset()
+        assert model_zero_init.get_states() == STATES_ZERO
 
     @pytest.mark.parametrize("states", [STATES_ZERO, STATES_NON_ZERO])
-    def test_reset_with_arg(self, nlm_zero_init, states):
-        nlm_zero_init.reset(states)
-        assert nlm_zero_init.get_states() == states
+    def test_reset_with_arg(self, model_zero_init, states):
+        model_zero_init.reset(states)
+        assert model_zero_init.get_states() == states
